https://trello.com/c/QyDhXpxX/995-rf-aitemtest-class-newitemtest-class
all tests has grouped according to type
AItemTest class is removed
