+++ src/test/java/school/redrover/NewItemTest.java
@@ -28,6 +28,19 @@ public class NewItemTest extends BaseTest {
     }
 
     @Test
+    public void testCreateNewFolder() {
+        List<String> itemsList = new HomePage(getDriver())
+                .clickNewItem()
+                .setItemName("Name")
+                .selectFolderAndClickOk()
+                .clickSaveButton()
+                .clickLogo()
+                .getItemList();
+
+        Assert.assertListContainsObject(itemsList, "Name", "Item not found");
+    }
+
+    @Test
     public void testCreateItemWithoutSelectedItemType() {
         boolean okButtonIsEnabled = new HomePage(getDriver())
                 .clickNewItem()
@@ -35,6 +48,24 @@ public class NewItemTest extends BaseTest {
                 .isOkButtonEnabled();
 
         Assert.assertFalse(okButtonIsEnabled);
+    }
+
+    @Test
+    public void testRenameFolder() {
+        List<String> itemsList = new HomePage(getDriver())
+                .clickNewItem()
+                .setItemName("Name")
+                .selectFolderAndClickOk()
+                .clickSaveButton()
+                .clickLogo()
+                .openItemDropdownWithSelenium("Name")
+                .renameFolderFromDropdown()
+                .setNewName("New Name")
+                .clickRename()
+                .clickLogo()
+                .getItemList();
+
+        Assert.assertListContainsObject(itemsList, "New Name", "Item not found");
     }
 
     @Test
@@ -76,6 +107,39 @@ public class NewItemTest extends BaseTest {
         Assert.assertFalse(IsOkButtonEnabled);
         Assert.assertEquals(validationMessage, hintTextWhenEmptyName);
         Assert.assertEquals(validationMessageColor, hintColor);
+    }
+
+    @Test
+    public void testCreateItemEmptyNameNegative() {
+        CreateNewItemPage createNewItemPage = new HomePage(getDriver())
+                .clickNewItem()
+                .setItemName("Name")
+                .clearItemNameField()
+                .selectFolder();
+
+        Assert.assertFalse(createNewItemPage.isOkButtonEnabled());
+    }
+
+    @Test
+    public void testCreateMulticonfigurationProject() {
+        List<String> itemsList = new HomePage(getDriver())
+                .clickNewItem()
+                .setItemName("Name")
+                .selectMultiConfigurationAndClickOk()
+                .clickSaveButton()
+                .clickLogo()
+                .getItemList();
+
+        Assert.assertListContainsObject(itemsList, "Name", "Item not found");
+    }
+
+    @Test
+    public void testCreateMulticonfigurationProjectNegative() {
+        CreateNewItemPage createNewItemPage = new HomePage(getDriver())
+                .clickNewItem()
+                .selectMultiConfiguration();
+
+        Assert.assertEquals(createNewItemPage.getItemNameHintText(), "» This field cannot be empty, please enter a valid name");
     }
 
     @Test(dependsOnMethods = "testDropdownNamesMenuContentWhenCopyProject")
